+++ server/src/config.ts
@@ -1,5 +1,23 @@
 import { Request } from "express";
 import type { Knex } from "knex";
+import { createLogger, Logger, transports, format } from "winston";
+const { combine, timestamp, splat, printf, label } = format;
+
+const level = process.env.LOG_LEVEL || "info";
+
+const timestampPrefixLogFormat = printf(
+  ({ level, message, label, timestamp }) => {
+    return `${timestamp} [${label}] ${level}: ${message}`;
+  }
+);
+
+const customFormatWithTimestamp = combine(
+  label({ label: "univaf" }),
+  timestamp(),
+  splat(),
+  timestampPrefixLogFormat
+);
+
 
 export function getApiKeys(): Array<string> {
   let keyList = process.env.API_KEYS;
@@ -35,3 +53,10 @@ export function loadDbConfig(): Knex.Config {
   const nodeEnv = process.env.NODE_ENV || "development";
   return knexfile[nodeEnv];
 }
+
+
+export const logger: Logger = createLogger({
+  format: customFormatWithTimestamp,
+  level: level,
+  transports: [new transports.Console()],
+});
